Added files via upload
main.py requests for you to input any number of digits as a phone number and then outputs matching words that match in words_filtered.txt

main_takeinputfile.py will take an input file containing phone numbers and output words that match each of the numbers in the phone number list. As of this initial commit, main_takeinputfile.py does not work yet.